refactor: Move writer generation to a separate class
