+++ Assets/Mirror/Editor/Weaver/Readers.cs
@@ -1,0 +1,7 @@
+using System;
+public class Readers
+{
+    public Readers()
+    {
+    }
+}
