+++ Assets/Mirror/Editor/Weaver/Writers.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a90060ad76ea044aba613080dd922709
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
